refactor(backend): split variant_validator_endpoint.py into sub-package
Decomposes the 702-LOC flat variant_validator_endpoint.py into a
sub-package at the same import path:

  variant_validator_endpoint/
    __init__.py          (34)   aggregator router: mounts /api/v2/variants
    schemas.py           (41)   4 Pydantic request/response models
    helpers.py           (25)   detect_notation_type
    validate_route.py    (156)  POST /validate
    annotate_route.py    (160)  POST /annotate
    recode_route.py      (190)  POST /recode + POST /recode/batch
    suggest_route.py     (91)   GET  /suggest/{partial_notation}

Every file is under 500 LOC. The aggregator router in __init__.py
preserves the /api/v2/variants prefix so app.main's
app.include_router(variant_validator_endpoint.router) call stays
unchanged.

HTTP surface byte-identical to the pre-Wave-4 flat module.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
